Fixes and updates
I fixed the camera/movement issue for real this time. Tank drive and
arcade drive are selectable by noting and denoting the respective drive
codes. I hope to eventually have it on a toggle button, but it doesn't
seem necessary, as whoever is driving will select their preferred drive
type so it hypothetically won't need to be changed frequently.