Gyroscope Update
Added syntax for the gyroscope. Please note that when increasing the
speed of the robot to -0.5, the robot has the tendency to drift to the
side. The robot does not seem to drift at -0.25. I'm guessing that it's
a problem with the floor that we are using, but I could be wrong. Either
way, assuming we have autonomous code for competition, I don't think
speed is the key element.