Update
Got touch sensors working for pulling the arm up. I left early, but as
far as I know, the touch sensor for bringing the arm down still needs to
be installed. Programming that will be a piece of cake. Please note the
current autonomous state when testing. Be sure to change it before
activating it, as the current mode needs serious work. We'll get it. I
believe in us.

- Brad Konsela