Update
This is the final update on bag-and-tag day.
I don't say this often enough, but I'm proud of all of you. Not just the
programmers either (even though we are the best). We have really done so
well through all of this. We have had to deal with a lot of challenges,
but we have managed to overcome them like champs.

- Brad Konsela