Regional Competition
Listen, guys:
I'm so proud of all of you. Seriously. Not sure who reads these little
blurbs of mine, but that doesn't matter. Thank all of you for everything
that you do.

Not only did we do an amazing job at regionals, we were also awarded
best rookie team, top seeded rookie team, and the rookie all star award.
These things don't come easy and don't come without a lot of teamwork,
patience, sweat, tears, and more sweat.

This being said, the fun (and the challenge) is not over yet. We have
several things that we need to work on before the national competition.
The biggest thing that I would like for us to accomplish
programmatically is command-based programming. I think that we are more
than capable of making it work. I plan on making a template and copying
the code and basically making an exact replica of our working
programming, only in the format of commands. Tough job, but we're a
tough team.

Good luck, great job, I'm proud of all of you.

- Brad Konsela