Final Worlds Update
Changelog:
* updated version label
* moved noted code to a notes section
* formatted the motor controller information
* polished and cleaned up code

Listen. Nobody reads these, I'm convinced of it, but that doesn't
matter. Maybe in years to come, some curious little programmer will
delve into this and look at these archives. And maybe they'll look back
at the 2016 challenge and think that this game was nothing, that it
wasn't a big deal, but they'll be wrong. This is a huge deal. I don't
need to even say that.

Is our robot simple? Yeah, it is. But that works for us. So. Well. We
know what we do and we did it.
Is our robot reliable? Kind of :) Sure, we had a few problems but we
located and fixed them quickly and effectively. Good for us.

Bradley will be the first one to tell anybody that it isn't the rank
that matters, but 16th is phenomenal. Out of 75 teams in our division,
600 teams total, we held our own in this competition. 16th is simply
phenomenal.

I said it in the final regionals update, and I'll say it again: I am so
proud of all of you. I truly am. Our rookie year, making it to Worlds,
simply extraordinary.

I can remember back to build season, week 0 competition, the first day
of regionals, how many times that it was so easy to think that we
couldn't do it or that our simplicity wouldn't work for us, but it did.

So again, I'm so proud of all of you. This is bittersweet for me because
I am leaving for college. I've driven the robot in competition for the
last time. I am honored to have had the chance.

Oh wait, you may be thinking that you won't be seeing much of me from
here on out, but you can't get rid of me that easily. I plan on coming
back as a mentor. I'm going to help programmers and drivers, so that we
can have similar success in years to come.
Thanks for everything.

-Brad Konsela